+++ src/app/modules/core/components/goals/goals.component.ts
@@ -5,7 +5,4 @@ import { Component } from '@angular/core';
   templateUrl: './goals.component.html',
   styleUrls: ['./goals.component.scss'],
 })
-export class GoalsComponent {
-  imageUrl = '../../assets/images/banner-images/features.svg';
-  constructor() {}
-}
+export class GoalsComponent {}