+++ src/app/modules/core/components/price-banner/price-banner.component.ts
@@ -1,6 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
-import { EntryPointService } from '../../services/entry-point.service';
+import { Component, Input } from '@angular/core';
 
 @Component({
   selector: 'app-price-banner',
@@ -10,6 +8,5 @@ import { EntryPointService } from '../../services/entry-point.service';
 export class PriceBannerComponent {
   @Input() textBanner: string;
   @Input() titleBanner: string;
-
-  constructor() {}
+  @Input() differentHeight: boolean;
 }