+++ src/app/modules/core/components/svg-components/angular-svg/angular-svg.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-angular-svg',
-  templateUrl: './angular.svg',
-})
-export class AngularSvgComponent {}