+++ src/app/modules/core/components/svg-components/ionic-svg/ionic-svg.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-ionic-svg',
-  templateUrl: './ionic.svg',
-})
-export class IonicSvgComponent {}