+++ src/app/modules/core/components/work-with-us/work-with-us.component.ts
@@ -7,5 +7,4 @@ import { Component } from '@angular/core';
 })
 export class WorkWithUsComponent {
   imageUrl = '../../assets/images/banner-images/features.svg';
-  constructor() {}
 }