+++ src/app/modules/core/core.module.ts
@@ -21,19 +21,9 @@ import { GoalsComponent } from './components/goals/goals.component';
 import { HeaderComponent } from './components/header/header.component';
 import { HelperDialogComponent } from './components/helper-dialog/helper-dialog.component';
 import { InfoCardComponent } from './components/info-card/info-card.component';
-import { ProcessInterestedComponent } from './components/process-interested/process-interested.component';
 import { SolutionDialogComponent } from './components/solution-dialog/solution-dialog.component';
 import { StepsComponent } from './components/steps/steps.component';
-import { AngularSvgComponent } from './components/svg-components/angular-svg/angular-svg.component';
-import { BootstrapSvgComponent } from './components/svg-components/bootstrap-svg/bootstrap-svg.component';
-import { CssSvgComponent } from './components/svg-components/css-svg/css-svg.component';
-import { HtmlSvgComponent } from './components/svg-components/html-svg/html-svg.component';
-import { IonicSvgComponent } from './components/svg-components/ionic-svg/ionic-svg.component';
-import { JavascriptSvgComponent } from './components/svg-components/javascript-svg/javascript-svg.component';
-import { NestSvgComponent } from './components/svg-components/nest-svg/nest-svg.component';
-import { TypescriptSvgComponent } from './components/svg-components/typescript-svg/typescript-svg.component';
 import { TextImageListComponent } from './components/text-image-list/text-image-list.component';
-import { TutorCardsComponent } from './components/tutor-cards/tutor-cards.component';
 import { UnderlineAnimationComponent } from './components/underline-animation/underline-animation.component';
 import { WorkWithUsComponent } from './components/work-with-us/work-with-us.component';
 import { ContactFormComponent } from './components/contact-form/contact-form.component';
@@ -41,6 +31,7 @@ import { MailSentDialogComponent } from './components/mail-sent-dialog/mail-sent
 import { PriceBannerComponent } from './components/price-banner/price-banner.component';
 import { InquiryComponent } from './components/inquiry/inquiry.component';
 import { PriceCardComponent } from './components/price-card/price-card.component';
+import { SponsorsCardComponent } from './components/sponsors-card/sponsors-card.component';
 
 @NgModule({
   imports: [
@@ -69,12 +60,12 @@ import { PriceCardComponent } from './components/price-card/price-card.component
     AcademyPictureComponent,
     AcademyBannerComponent,
     GoalsComponent,
-    ProcessInterestedComponent,
     WorkWithUsComponent,
     UnderlineAnimationComponent,
     ContactFormComponent,
     PriceBannerComponent,
     InquiryComponent,
+    SponsorsCardComponent,
 
     // Dialogs
     HelperDialogComponent,
@@ -84,15 +75,6 @@ import { PriceCardComponent } from './components/price-card/price-card.component
     SolutionDialogComponent,
     VersionHistoryComponent,
 
-    // SVGs
-    HtmlSvgComponent,
-    CssSvgComponent,
-    JavascriptSvgComponent,
-    TypescriptSvgComponent,
-    AngularSvgComponent,
-    BootstrapSvgComponent,
-    IonicSvgComponent,
-    NestSvgComponent,
     ContactCardComponent,
     FooterComponent,
     HeaderComponent,
@@ -102,7 +84,6 @@ import { PriceCardComponent } from './components/price-card/price-card.component
     DialogComponent,
     StepsComponent,
     FeaturesComponent,
-    TutorCardsComponent,
     CompanyTestimonialComponent,
     MailSentDialogComponent,
   ],
@@ -116,27 +97,18 @@ import { PriceCardComponent } from './components/price-card/price-card.component
     FooterComponent,
     HeaderComponent,
     BrandBannerComponent,
-    HtmlSvgComponent,
-    CssSvgComponent,
-    JavascriptSvgComponent,
-    TypescriptSvgComponent,
-    AngularSvgComponent,
-    BootstrapSvgComponent,
-    IonicSvgComponent,
-    NestSvgComponent,
     ContactCardComponent,
     StepsComponent,
     FeaturesComponent,
-    TutorCardsComponent,
     CompanyTestimonialComponent,
     AcademyPictureComponent,
     AcademyBannerComponent,
     GoalsComponent,
-    ProcessInterestedComponent,
     WorkWithUsComponent,
     ContactFormComponent,
     PriceBannerComponent,
     InquiryComponent,
+    SponsorsCardComponent,
   ],
 })
 export class CoreModule {}