+++ src/app/modules/core/interfaces/sponsor.interface.ts
@@ -1,0 +1,10 @@
+export interface Sponsor {
+  imageUrl: string;
+  imageAlt: string;
+  title: string;
+  careerType: string[];
+  careerLink: string;
+  mainPageLink: string;
+  statement: string;
+  isOnlySponsor: boolean;
+}