+++ src/app/sections/academy/sections/packages/packages.component.ts
@@ -5,6 +5,4 @@ import { Component } from '@angular/core';
   templateUrl: './packages.component.html',
   styleUrls: ['./packages.component.scss'],
 })
-export class PackagesComponent {
-  constructor() {}
-}
+export class PackagesComponent {}