+++ src/app/sections/start/start.module.ts
@@ -14,6 +14,8 @@ import { ContactPageComponent } from '../landing-pages/contact-page/contact-page
 import { SubscriptionPageComponent } from '../landing-pages/subscription-page/subscription-page.component';
 import { TeamPageComponent } from '../landing-pages/team-page/team-page.component';
 import { WorkshopsComponent } from '../landing-pages/workshops/workshops.component';
+import { SponsorsPageComponent } from '../landing-pages/sponsors-page/sponsors-page.component';
+import { TypePipe } from './pipes/type.pipe';
 
 @NgModule({
   declarations: [
@@ -26,6 +28,8 @@ import { WorkshopsComponent } from '../landing-pages/workshops/workshops.compone
     SubscriptionPageComponent,
     TeamPageComponent,
     WorkshopsComponent,
+    SponsorsPageComponent,
+    TypePipe,
   ],
   imports: [
     StartRoutingModule,